Added a toggle for the menu.
